Move monitoringServices to plugins directory
