+++ packages/lambda/src/api/getPublicMetrics/index.js
@@ -1,5 +1,5 @@
 import MetricsStore from 'db/metrics'
-import 'model/monitoringServices'  // load monitoring services
+import 'plugins/monitoringServices'  // load monitoring services
 import { metricStatusVisible } from 'utils/const'
 
 export async function handle (event, context, callback) {
