+++ packages/lambda/src/plugins/monitoringServices/self.js
